+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [7.8.1] - 2026-01-14
+
+### Added
+
+- Debilitating Strike/Double Debilitation
+- Terrified Retreat
+
 ## [7.8.0] - 2026-01-13
 
 PF2e System v7.8.0 Update
@@ -399,7 +406,8 @@ I don't remember all the changes that I made before I stopped working on this fo
 - Auto Self-Applied Effects (This is disabled if PF2e Toolbelt is doing this as well)
 - Swashbuckler Panache
 
-[Unreleased]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.8.0...HEAD
+[Unreleased]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.8.1...HEAD
+[7.8.1]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.8.0...v7.8.1
 [7.8.0]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.2.0...v7.8.0
 [7.2.0]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.0.1...v7.2.0
 [7.0.1]: https://github.com/7H3LaughingMan/pf2e-assistant/compare/v7.0.0...v7.0.1